pyproject.toml
[project]
name = "rag-telegram-bot"
version = "0.1.0"
description = "Telegram bot with RAG (Retrieval-Augmented Generation) for answering questions based on knowledge base"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.3.25",
    "langchain-gigachat>=0.3.10",
    "langchain-community>=0.3.0",
    "langchain-huggingface>=0.0.1",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.4",
    "tqdm>=4.66.0",
    "python-telegram-bot>=20.0",
]
